+++ lib/screens/account_tab/usernames/username_detailed_screen.dart
@@ -1,6 +1,6 @@
 import 'package:animations/animations.dart';
 import 'package:anonaddy/models/username/username_model.dart';
-import 'package:anonaddy/screens/account_tab/username_default_recipient.dart';
+import 'package:anonaddy/screens/account_tab/usernames/username_default_recipient.dart';
 import 'package:anonaddy/shared_components/alias_created_at_widget.dart';
 import 'package:anonaddy/shared_components/bottom_sheet_header.dart';
 import 'package:anonaddy/shared_components/constants/material_constants.dart';
@@ -17,7 +17,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../global_providers.dart';
+import '../../../global_providers.dart';
 
 class UsernameDetailedScreen extends ConsumerWidget {
   @override
