+++ lib/screens/home_screen.dart
@@ -1,7 +1,5 @@
 import 'package:anonaddy/global_providers.dart';
-import 'package:anonaddy/screens/failed_deliveries/failed_deliveries_screen.dart';
 import 'package:anonaddy/screens/search_tab/search_tab.dart';
-import 'package:anonaddy/screens/settings_screen/settings_screen.dart';
 import 'package:anonaddy/services/data_storage/search_history_storage.dart';
 import 'package:anonaddy/shared_components/constants/material_constants.dart';
 import 'package:anonaddy/shared_components/constants/ui_strings.dart';
@@ -17,7 +15,9 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'account_tab/account_tab.dart';
 import 'alias_tab/alias_tab.dart';
-import 'alias_tab/create_new_alias/create_new_alias.dart';
+import 'home_screen_componenets/create_new_alias/create_new_alias.dart';
+import 'home_screen_componenets/failed_deliveries/failed_deliveries_screen.dart';
+import 'home_screen_componenets/settings_screen/settings_screen.dart';
 
 class HomeScreen extends StatefulWidget {
   @override
@@ -172,6 +172,7 @@ class _HomeScreenState extends State<HomeScreen> {
     );
   }
 
+  //todo extract into a widget that goes into home_screen_componeents
   Future buildUpdateNews(BuildContext context) {
     final size = MediaQuery.of(context).size;
